requirements.txt
# Django
django==6.0
django-liveview==2.1.9

# Django Server
daphne==4.2.0

# Django Channels
channels==4.2.0

# Redis
redis==5.2.0
channels_redis==4.2.1

# Testing
pytest==8.3.4
pytest-django==4.9.0
